refactor(kb): move readSidecar to src/cli/utils/sidecar.ts
Extract SidecarData interface and readSidecar function from kb.ts into
a dedicated utility module. kb.ts re-exports both for backward compatibility
so existing test imports from kb.js continue to work unchanged.
